added skip and limit via findBy
